making computing coverage more robust.
Sometimes there are two UniProt IDs map the same gene. If fasta file needs to be presented at gene level one way to resolve ambiguity is to select longest sequence. This can lead to troubles when there was a peptide matching to shorter protein product of the gene. Example is MOCS2 gene and SETISVPQEIK peptide.